+++ api/fe.js
@@ -1,0 +1,7 @@
+const { generateModelScript } = require('../forward_engineering/helpers/generateModelScript');
+const { validate } = require('../forward_engineering/helpers/validation/validate');
+
+module.exports = {
+	generateModelScript,
+	validate,
+};
